+++ src/helpers/index.js
@@ -76,6 +76,8 @@ export {
   normalizeLanguageName,
   hasApiExpander,
 } from '@plone/volto/helpers/Utils/Utils';
+
+export { getImageAttributes } from './Image/Image';
 export { messages } from './MessageLabels/MessageLabels';
 export {
   withBlockSchemaEnhancer,